+++ R/settings.meta-internal.R
@@ -19,8 +19,10 @@ catarg <- function(x, newline = TRUE, end = "") {
 specificSettings <- function(args, new, setting, quietly = FALSE) {
   isnull.old <- as.vector(unlist(lapply(.settings[args], is.null)))
   ischar.old <- as.vector(unlist(lapply(.settings[args], is.character)))
-  old <- as.vector(unlist(.settings[args]))
+  old <- rep("character", length(isnull.old))
+  old[!isnull.old] <- as.vector(unlist(.settings[args]))
   ##
+  isnull.new <- as.vector(unlist(lapply(new, is.null)))
   ischar.new <- as.vector(unlist(lapply(new, is.character)))
   new <- as.vector(unlist(new))
   ##